feat(inputs): get user inputs
Issue #3 get the users input.

I call in 'main()' the new implemented function 'kairos_get_input()' that return a vector
with the user inputs.

First you get the project name which will be the project repository and the project binary
name, after you get the 'HTTPS Github URL', the string is a '\n' if the user don't want to
clone a project. Finally we get the coding language, you can have two anser for each
language, the name or the index.

We push all the inputs in a vector and we return it.

Deniz Uzuntok Husson
deniz.uzuntok-husson@epitech.eu